Move Org models to models.py
