+++ backend/btrixcloud/profiles.py
@@ -11,12 +11,11 @@ from fastapi import APIRouter, Depends, Request, HTTPException
 from pydantic import BaseModel, UUID4, HttpUrl
 import aiohttp
 
-from .orgs import Organization
 from .pagination import DEFAULT_PAGE_SIZE, paginated_format
 from .users import User
 
 from .db import BaseMongoModel
-from .models import CrawlConfigIdNameOut
+from .models import CrawlConfigIdNameOut, Organization
 
 
 BROWSER_EXPIRE = 300
